Remove clean command from autogen.sh
The user can use `git clean -fxd -'**/Makefile'` to achieve the same
with more accuracy and no need to maintain a list of static files in any
place in the project.

The same can be done with a target like `clean-local: @-rm -rfdv ...` in
makes but is it really even needed? Discussion is needed to add this
back in one way or another.